added edit news and hide/show recipes

+++ model/NewsModel.php
@@ -22,10 +22,10 @@ class NewsModel extends Model
         $pdo = parent::connexion();
 
         $qtf = $pdo->prepare(
-            "SELECT `idNews`, `titre`, `corps`, `image`, `video`, `date` FROM news
-            WHERE `idNews` = :idNews"
+            "SELECT idNews, titre, corps, image, video, date FROM news
+            WHERE idNews = :idNews"
         );
-        $qtf->bindParam(':idNews', $idNews, PDO::PARAM_INT);
+        $qtf->bindParam(':idNews', $idNews);
         $qtf->execute();
         $result = $qtf->fetch();
 
@@ -33,7 +33,8 @@ class NewsModel extends Model
         return $result;
     }
 
-    public function ajouterNews($titre, $corps, $image, $video){
+    public function ajouterNews($titre, $corps, $image, $video)
+    {
 
         $imageLink = parent::ajouterImage($image);
         if (empty($imageLink)) {
@@ -56,11 +57,74 @@ class NewsModel extends Model
         $qtf->bindParam(':image', $imageLink);
         $qtf->bindParam(':video', $videoLink);
         $qtf->execute();
-        
+
         parent::deconnexion($pdo);
     }
 
-    public function deleteNews($idNews){
+    public function modifierNews($idNews, $titre, $corps, $image, $video)
+    {
+        $news = $this->getNewsById($idNews);
+        if (!$news) {
+            return;
+        }
+
+        $imageLink = $news["image"];
+        $videoLink = $news["video"];
+
+        if (isset($image) && !empty($image)) {
+            if (!empty($imageLink))
+                unlink(__DIR__ . '/..' . $imageLink);
+            $tmpLink = parent::ajouterImage($image);
+            if (!empty($tmpLink))
+                $imageLink = $tmpLink;
+        }
+
+
+        if (isset($video) && !empty($video)) {
+            if (!empty($videoLink))
+                unlink(__DIR__ . '/..' . $videoLink);
+            $tmpLink = parent::ajouterVideo($video);
+            if (!empty($tmpLink))
+                $videoLink = $tmpLink;
+        }
+
+        $pdo = parent::connexion();
+        $qtf = $pdo->prepare(
+            "UPDATE news
+            SET titre = :titre,
+            corps = :corps,
+            image= :image,
+            video= :video
+            WHERE idNews = :idNews;"
+        );
+        $qtf->bindParam(':titre', $titre);
+        $qtf->bindParam(':corps', $corps);
+        $qtf->bindParam(':image', $imageLink);
+        $qtf->bindParam(':video', $videoLink);
+        $qtf->bindParam(':idNews', $idNews);
+        $qtf->execute();
+
+        parent::deconnexion($pdo);
+    }
+
+    public function deleteNews($idNews)
+    {
+
+        $news = $this->getNewsById($idNews);
+        if (!$news) {
+            return;
+        }
+
+        $imageLink = $news["image"];
+        $videoLink = $news["video"];
+
+        if (!empty($imageLink))
+            unlink(__DIR__ . '/..' . $imageLink);
+
+
+        if (!empty($videoLink))
+            unlink(__DIR__ . '/..' . $videoLink);
+
         $pdo = parent::connexion();
 
         $qtf = $pdo->prepare("DELETE FROM `news` WHERE idNews = :idNews");
